apps/vendas/serializers.py
from rest_framework import serializers
from apps.vendas.models import Vendedor, Venda

class VendedorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vendedor
        fields = '__all__'

class VendaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Venda
        fields = '__all__'

