pyproject.toml
[tool.poetry]
name = "desafio-corpsystem"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
requests = "^2.32.3"
django-mysql = "^4.14.0"
mysqlclient = "^1.2.5"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
